requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
tiktoken>=0.5.0
scikit-learn>=1.3.0
reportlab>=4.0.0